UPDATE - 문의게시판

+++ ToyMarket/src/main/java/com/toymarket/vo/Products.java
@@ -9,7 +9,7 @@ public class Products {
 	private String brand;
 	private String name;
 	private int price;
-	private int discountRate;
+	private double discountRate;
 	private String discountYN;
 	private int likeCount;
 	private String sellUnit;
@@ -64,11 +64,11 @@ public class Products {
 		this.price = price;
 	}
 
-	public int getDiscountRate() {
+	public double getDiscountRate() {
 		return discountRate;
 	}
 
-	public void setDiscountRate(int discountRate) {
+	public void setDiscountRate(double discountRate) {
 		this.discountRate = discountRate;
 	}
